Address review: centralize error handling, move domain errors back, add From impls
- Move StatusListError back to src/web/handlers/status_list/error.rs
  (domain errors should not live in the API error module)
- Remove re-export-only error.rs module; make it pub(crate)
- Add From<RepositoryError> for ApiError with logging, eliminating
  manual ApiError::internal() mapping in handlers
- Move From<CredentialError> for ApiError to errors.rs for centralized
  error conversion (per cloud-identity-wallet pattern)
- Use idiomatic ? propagation in handlers instead of manual map_err
- Remove #[allow(dead_code)] from get_error_message (method is used)
- Remove .markdownlint.json ignores for .kilo/ and node_modules/
- Move integration test file to unit tests in respective source modules
