+++ .markdownlint.json
@@ -3,6 +3,5 @@
   "MD022": false,
   "MD024": false,
   "MD032": false,
-  "MD060": { "style": "aligned" },
-  "ignores": [".kilo/", "node_modules/"]
+  "MD060": { "style": "aligned" }
 }
